+++ otterdog/models/repository.py
@@ -35,6 +35,7 @@ class Repository(ModelObject):
     has_wiki: bool
     is_template: bool
     template_repository: str = dataclasses.field(metadata={"read_only": True})
+    post_process_template_content: list[str] = dataclasses.field(metadata={"read_only": True})
     topics: list[str]
     default_branch: str
     allow_rebase_merge: bool